requirements.txt
pandas==1.3.4
requests==2.26.0
pymongo==4.0.1
python-dotenv==0.19.2
